Moving main_class and java_opts to configs
As a result of https://review.openstack.org/#/c/69982 main_class
and java_opts are now expected to be part of the job configs in
the form of edp.java.main_class and edp.java.java_opts.  The dashboard
masks this fact from the user by still presenting those as separate
boxes on the job configuration page and doing the translation
on their bahalf.

Partial-implements: blueprint edp-config-parameter-prefix

Change-Id: I923af44893d290a21a9644d3bda1324fc9060478