Add sahara_dashboard to ADD_INSTALLED_APPS for translation support
The infra script assumes that transaltion files is placed under
sahara_dashboard/locale and it is required that 'sahara_dashboard'
is included in ADD_INSTALLED_APPS to make translation work.

It did not work to add symlink to 'sahara_dashboard/locale'
and it was reverted by commit 55512fcc5e8d2db09896bcc55352912f8e650b08.
This commit explores another way to support translation.

Note that other entries in ADD_INSTALLED_APPS are not touched
to avoid breaking other stuff of sahara-dashboard.
I don't know the background of using the current ADD_INSTALLED_APPS,
so cleaning them up is out of scope of this patch.

Change-Id: I214fca300df771ff3f35e8a9efe216104f48ebd4