Ensure to compile translation catalogs
A working directory in DevStack is not sahara-dashboard directory,
so the directory path must be specified by an absolute path.

In addition, django manage.py works if 'locale' directory exists.
This commit uses django manage.py to simplify the command.

When I tested the devstack plugin for translation previously,
it seems compiled message catalogs existed and I didn't notice it.

Change-Id: I8dde297976faf01a6dbe1f70ec5cd3be4ba3f160