Use %s when non-ascii characters can be contained
.format() in python 2.7 cannot handle a unicode string if it contains
non-ASCII characters. This breaks i18n situations.

The behavior of .format() is different between python2 and python3.
We should use .format() only when we are confident that .format()
can handle passed values (like ASCII string or integer) correctly.

This commit does not touch other usages of .format() which work well.

Change-Id: I329b33c1c7776d7f24488572107f7ea35f418398
Closes-Bug: #1551014