Hiding undesired nav-pills
Recently, nav-tabs became nav-pills.  That lead to them
showing-up in places we were previously hiding the
nav-tabs.  This change hides all pill-tabs in places where
we'd rather not see them.  It also refactors the hiding
a bit.  Eventually, we will need a common library for
shared scripts between jobs and custers, but this
addresses the bug at hand for now.  Also updating the
datasourcepage so that it no longer expects a tabbed
form (since the only "tab" that was there is now hidden).

Change-Id: I31dd0c402580d5414769bddecdc077b01eac4379
Closes-Bug:  #1566396