Convert to unicode explicitly to avoid UnicodeDecodeError
In a zh-cn environment in horizon, when click the button of updating
plugin, chinese characters can lead to error.
Because the type of 'prefix_trans' is non unicode but 'name' is unicode.
Python will decode 'prefix_trans' firstly. If 'prefix_trans' contains
chinese characters then will lead to errors saying out of range 128.
So we convert 'prefix_trans' to unicode explicitly and python will not
decode because both 'prefix_trans' and 'name' is unicode. In fact,
'prefix_trans', 'prefix' and 'name' will be unicode at last.

Change-Id: I75f3334f7e6bac9dc2819ac2632636ece95efc31
Closes-Bug:#1622864