Rename the 'SAHARA_AUTO_IP_ALLOCATION_ENABLED' config parameter
The 'SAHARA_AUTO_IP_ALLOCATION_ENABLED' config parameter in Sahara
is confusing. If you wish to disable floating IP options during node
group template creation, set it 'True', but this name sounds like it
has to do with automatically associating a floating IP. So this patch
renames it to a meaningful variable name.

Change-Id: If8f71ea21cb71bbffe996f7db88b7683ad91ecac