Fix attrs list of LabeledInput.render() method
The item (and the buttom to create a new job binary) didn't work anymore
in Django 1.11, due to a change in the parameter list of the render()
method: it was previously based on the position, now the arguments are
explicitly passed using their names.

Thanks to https://github.com/springload/wagtaildraftail/issues/51
for the explanation.

Closes-Bug: #1743384
Change-Id: I241021c593ad50591f0155dfb364afbc662a4794