Require the (soon-to-be) Rocky version of Horizon
... or at least the first published tarball which will lead to Rocky.
On the other side this means forcing the usage of mox as long as
there are tests that depends on it (the switch was already flipped
in Horizon, so it needs to be enabled here).

This allows us to depend on newer Horizon features, including changes
to the test suite that are useful for the mox removal effort.

Change-Id: Ibf95b642f49aee9b3165f46e0422ff6bf56e9391