Fix the coverage tox target
It was broken for a long while, since the migration
of the dashboard code into Horizon, and then back
into this repository, as .testr.conf was removed.

Use coverage directly instead of testrepository,
the same way as Horizon does. We should port
the normal test runner to the Django test runner too,
instead of the custom run_tests.sh script, but
that's for another (future) review.

Change-Id: I08966ddc7e891156a1c934bb115ce8a02f77ec83