Remove the legacy integration tests job
The tests should be fixed in master/stein first, because the integration
test frameworks is available from horizon only for those releases.
As the frameworks is unlikely to be backported, there is no point
in keeping a broken version of a native job.

If the integration tests are fixed at some point, this decision can be
re-evaluated.

This was discussed during the Sahara weekly meeting:
http://eavesdrop.openstack.org/meetings/sahara/2019/sahara.2019-02-28-14.01.log.html#l-109

Change-Id: I311203779bd3e6c6660cfe1b84e89c7dc0825c98
(cherry picked from commit ea3a84ac4ef86e64891107c66d02459a87933568)