Overhaul of the unit test jobs and tox environments
Since Rocky, horizon depends on Django 2.0 when Python 3 is used.
The old sahara-dashboard-tox-py35dj20 and the py35dj20 tox environments
are thus no longer needed.
Add few more jobs:
- the py36 and py37 jobs and the related tox environments;
- a py3/django 1.11 job and tox environment, and set is a voting
  like its Horizon counterpart;
- the tox environment for a django 2.1 job (Horizon provides one as well.)

Finally, don't require the last version of horizon for the py3/django 1.11,
but relies on the stable version (which should provide the proper library
interface, as it happen with the other unit test jobs.)

Change-Id: Ic76b454f36d81a25b40e91a88d131f20a8df12f9