Fix (lower) requirements (sync with horizon constraints)
Mostly sync sphinx, but also other packages.

Also bump horizon to the first stable version
for the stein branch. The b2 release was intended
to be temporary to allow us to bump the dependency
as soon as possible, but there is no py3 wheel
apparently and the lower-constraint job fails.

As horizon is not tracked by requirements.git
limit its upper constraint to the highest version
for this branch to prevent failures due to changed APIs.

Change-Id: I4f6d26a374113dcbe1e52d981f59829616aa6a3a